Moved tests to lexer module file
