--- src/compiler/test/mod.rs
@@ -1,1 +1,0 @@
-mod test_lexer;
